--- src/BenchmarksApps/SslStream/Client/ClientOptions.cs
@@ -1,35 +1,0 @@
-using System.CommandLine;
-using System.CommandLine.Binding;
-using System.Security.Authentication;
-
-using ConnectedStreams.Client;
-using SslStreamCommon;
-
-namespace SslStreamClient;
-public class SslStreamClientOptions : ClientOptions, ISslOptions
-{
-    public bool AllowTlsResume { get; set; }
-    public SslProtocols EnabledSslProtocols { get; set; }
-
-    public override string ToString()
-        => $"{base.ToString()}, AllowTlsResume: {AllowTlsResume}, EnabledSslProtocols: {EnabledSslProtocols}";
-}
-
-public class SslStreamOptionsBinder : BinderBase<SslStreamClientOptions>
-{
-    public static void AddOptions(RootCommand command)
-    {
-        ClientOptionsBinder.AddOptions(command);
-        SslOptionsCommon.AddOptions(command);
-    }
-
-    protected override SslStreamClientOptions GetBoundValue(BindingContext bindingContext)
-    {
-        var options = new SslStreamClientOptions();
-
-        ClientOptionsBinder.BindOptions(options, bindingContext);
-        SslOptionsCommon.BindOptions(options, bindingContext);
-
-        return options;
-    }
-}
