+++ src/System/Net/Security/Quic/Server/Program.cs
@@ -1,0 +1,87 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.Net;
+using System.Net.Security;
+using System.CommandLine;
+
+using System.Net.Security.Benchmarks.Server;
+using System.Net.Quic;
+using Common;
+
+internal class Program
+{
+    private static async Task Main(string[] args)
+    {
+        var server = new QuicBenchmarkServer();
+        await server.RunCommandAsync<OptionsBinder>(args);
+    }
+}
+
+// The benchmarks are only run on Windows and Linux
+#pragma warning disable CA1416 // "This call site is reachable on all platforms. It is only supported on: 'linux', 'macOS/OSX', 'windows'."
+
+internal class QuicBenchmarkServer : SslBenchmarkServer<ServerOptions>
+{
+    public override string Name => "QUIC benchmark server";
+    public override string MetricPrefix => "quic";
+
+    public override void AddCommandLineOptions(RootCommand rootCommand)
+        => OptionsBinder.AddOptions(rootCommand);
+
+    public override bool IsExpectedException(Exception e)
+        => e is QuicException qe && qe.QuicError == QuicError.ConnectionAborted;
+
+    public override async Task<IBaseListener<IServerConnection>> ListenAsync(ServerOptions options, CancellationToken ct)
+    {
+        var sslOptions = CreateSslServerAuthenticationOptions(options);
+        var connectionOptions = new QuicServerConnectionOptions()
+        {
+            DefaultStreamErrorCode = 123,
+            DefaultCloseErrorCode = 456,
+            ServerAuthenticationOptions = sslOptions,
+            MaxInboundBidirectionalStreams = 20000
+        };
+        var listenerOptions = new QuicListenerOptions()
+        {
+            ListenEndPoint = new IPEndPoint(IPAddress.IPv6Any, options.Port),
+            ApplicationProtocols = connectionOptions.ServerAuthenticationOptions.ApplicationProtocols!,
+            ConnectionOptionsCallback = (_, _, _) => ValueTask.FromResult(connectionOptions)
+        };
+
+        return new QuicServerListener(await QuicListener.ListenAsync(listenerOptions, ct), options);
+    }
+}
+
+internal class QuicServerListener(QuicListener _listener, ServerOptions _serverOptions) : IListener
+{
+    public EndPoint LocalEndPoint => _listener.LocalEndPoint;
+
+    public async Task<IServerConnection> AcceptAsync(CancellationToken cancellationToken)
+        => new QuicServerConnection(await _listener.AcceptConnectionAsync(cancellationToken), _serverOptions);
+
+    public ValueTask DisposeAsync() => _listener.DisposeAsync();
+}
+
+internal class QuicServerConnection(QuicConnection _connection, ServerOptions _serverOptions) : IServerConnection
+{
+    public Task CompleteHandshakeAsync(CancellationToken _) => Task.CompletedTask;
+    public SslApplicationProtocol NegotiatedApplicationProtocol => _connection.NegotiatedApplicationProtocol;
+    public bool IsMultiplexed => true;
+    private static readonly byte[] s_byteBuf = new byte[1];
+
+    public async Task<Stream> AcceptInboundStreamAsync(CancellationToken cancellationToken)
+    {
+        var stream = await _connection.AcceptInboundStreamAsync(cancellationToken);
+        if (_serverOptions.ReceiveBufferSize == 0)
+        {
+            // drain the single byte used to open the stream
+            _ = await stream.ReadAsync(s_byteBuf, cancellationToken).ConfigureAwait(false);
+        }
+        return stream;
+    }
+
+    public ValueTask DisposeAsync() => _connection.DisposeAsync();
+}
+
+#pragma warning restore CA1416
