+++ src/System/Net/System.Net.sln
@@ -1,0 +1,70 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SslStreamClient", "Security\SslStream\Client\SslStreamClient.csproj", "{50BBC830-3EFB-4274-80BF-A247D5C1FF97}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SslStreamServer", "Security\SslStream\Server\SslStreamServer.csproj", "{2E42BCAC-7A6B-4E98-8463-51AF09B82763}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QuicClient", "Security\Quic\Client\QuicClient.csproj", "{95886C88-B124-40EE-83AE-434325586443}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "QuicServer", "Security\Quic\Server\QuicServer.csproj", "{9D2CDEFA-99CA-4D51-954A-3E385BB1498C}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{D7473ABA-AAA6-47E8-BDB4-FFDCFC538615}"
+	ProjectSection(SolutionItems) = preProject
+		..\..\..\build\dependencies.props = ..\..\..\build\dependencies.props
+		..\..\..\build\dependencies.targets = ..\..\..\build\dependencies.targets
+		Directory.Build.props = Directory.Build.props
+		..\..\Directory.Build.props = ..\..\Directory.Build.props
+		..\..\Directory.Build.targets = ..\..\Directory.Build.targets
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug_Database|Any CPU = Debug_Database|Any CPU
+		Debug|Any CPU = Debug|Any CPU
+		Release_Database|Any CPU = Release_Database|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{50BBC830-3EFB-4274-80BF-A247D5C1FF97}.Debug_Database|Any CPU.ActiveCfg = Debug_Database|Any CPU
+		{50BBC830-3EFB-4274-80BF-A247D5C1FF97}.Debug_Database|Any CPU.Build.0 = Debug_Database|Any CPU
+		{50BBC830-3EFB-4274-80BF-A247D5C1FF97}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{50BBC830-3EFB-4274-80BF-A247D5C1FF97}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{50BBC830-3EFB-4274-80BF-A247D5C1FF97}.Release_Database|Any CPU.ActiveCfg = Release_Database|Any CPU
+		{50BBC830-3EFB-4274-80BF-A247D5C1FF97}.Release_Database|Any CPU.Build.0 = Release_Database|Any CPU
+		{50BBC830-3EFB-4274-80BF-A247D5C1FF97}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{50BBC830-3EFB-4274-80BF-A247D5C1FF97}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2E42BCAC-7A6B-4E98-8463-51AF09B82763}.Debug_Database|Any CPU.ActiveCfg = Debug_Database|Any CPU
+		{2E42BCAC-7A6B-4E98-8463-51AF09B82763}.Debug_Database|Any CPU.Build.0 = Debug_Database|Any CPU
+		{2E42BCAC-7A6B-4E98-8463-51AF09B82763}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2E42BCAC-7A6B-4E98-8463-51AF09B82763}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2E42BCAC-7A6B-4E98-8463-51AF09B82763}.Release_Database|Any CPU.ActiveCfg = Release_Database|Any CPU
+		{2E42BCAC-7A6B-4E98-8463-51AF09B82763}.Release_Database|Any CPU.Build.0 = Release_Database|Any CPU
+		{2E42BCAC-7A6B-4E98-8463-51AF09B82763}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2E42BCAC-7A6B-4E98-8463-51AF09B82763}.Release|Any CPU.Build.0 = Release|Any CPU
+		{95886C88-B124-40EE-83AE-434325586443}.Debug_Database|Any CPU.ActiveCfg = Debug_Database|Any CPU
+		{95886C88-B124-40EE-83AE-434325586443}.Debug_Database|Any CPU.Build.0 = Debug_Database|Any CPU
+		{95886C88-B124-40EE-83AE-434325586443}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{95886C88-B124-40EE-83AE-434325586443}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{95886C88-B124-40EE-83AE-434325586443}.Release_Database|Any CPU.ActiveCfg = Release_Database|Any CPU
+		{95886C88-B124-40EE-83AE-434325586443}.Release_Database|Any CPU.Build.0 = Release_Database|Any CPU
+		{95886C88-B124-40EE-83AE-434325586443}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{95886C88-B124-40EE-83AE-434325586443}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9D2CDEFA-99CA-4D51-954A-3E385BB1498C}.Debug_Database|Any CPU.ActiveCfg = Debug_Database|Any CPU
+		{9D2CDEFA-99CA-4D51-954A-3E385BB1498C}.Debug_Database|Any CPU.Build.0 = Debug_Database|Any CPU
+		{9D2CDEFA-99CA-4D51-954A-3E385BB1498C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9D2CDEFA-99CA-4D51-954A-3E385BB1498C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9D2CDEFA-99CA-4D51-954A-3E385BB1498C}.Release_Database|Any CPU.ActiveCfg = Release_Database|Any CPU
+		{9D2CDEFA-99CA-4D51-954A-3E385BB1498C}.Release_Database|Any CPU.Build.0 = Release_Database|Any CPU
+		{9D2CDEFA-99CA-4D51-954A-3E385BB1498C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9D2CDEFA-99CA-4D51-954A-3E385BB1498C}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {2D5D8564-77DF-4B39-8EC3-176ED3CCC754}
+	EndGlobalSection
+EndGlobal
